move blend-out logic to App.tsx
